backend/requirements.txt
# Framework web y extensiones
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.4
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.2
Flask-Mail==0.9.1
Flask-Limiter==3.3.1
python-dotenv==1.0.0

# Base de datos
SQLAlchemy==2.0.20
psycopg2-binary==2.9.7
alembic==1.11.3

# Utilidades
numpy==1.25.2
pandas==2.0.3
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0

# Manejo de datos
marshmallow==3.20.1
jsonschema==4.19.0

# Testing
pytest==7.4.0
pytest-flask==1.2.0
pytest-cov==4.1.0

# Desarrollo y seguridad
black==23.7.0
flake8==6.1.0
isort==5.12.0
passlib==1.7.4
bcrypt==4.0.1
